Done with Single Node
As it was discussed that the Scripts for setting up OpenStack
Training Labs will be one of the Doc Sprints Targets for this
week, I am done with the initial work on these scripts. They
need loads of testing and will probably not work as of now.

I suggest people to use Virtual Box unless you dont mind breaking
your machine badly ( :D ) and take a snapshot of your system before
messing around with it.